Remove NodeController
