+++ src/Sulu/Bundle/PageBundle/Resources/config/routing_api.yml
@@ -8,18 +8,6 @@ sulu_page.resource_locator:
     name_prefix: sulu_page.
     resource: sulu_page.resource_locator_controller
 
-sulu_page.node:
-    type: rest
-    name_prefix: sulu_page.
-    resource: Sulu\Bundle\PageBundle\Controller\NodeController
-
-sulu_page.post_node_trigger:
-    path: /nodes/{id}.{_format}
-    methods: POST
-    defaults:
-        _controller: Sulu\Bundle\PageBundle\Controller\NodeController::postTriggerAction
-        _format: json
-
 sulu_page.page:
     type: rest
     name_prefix: sulu_page.
